The code used to detect colors on a photo
*It is important to have SimpleCV Installed on the device the user is trying to run this script on. The code is originally executed using python script*

Don't change anything in this code without my consent, I can explain this to people if they are curious how the code works and what each line does. 

The above code can take a photo and run a series of tests to identify colors Blue, Green and Red on the same image. It will then create 3 separate photos highlighting the colors if successfully recognized. 

It is important to know that the success of the code is HIGHLY dependent on lighting environment around the photograph initially being taken. Will further explain and demonstrate in class on Monday!